@@ -4,7 +4,7 @@ version: 3.0.0
 homepage: https://www.moengage.com
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.12.0 <2.0.0"
 
 dependencies:
