@@ -2,10 +2,7 @@ class MoEGeoLocation {
   double latitude;
   double longitude;
 
-  MoEGeoLocation(double latitude, double longitude) {
-    this.latitude = latitude;
-    this.longitude = longitude;
-  }
+  MoEGeoLocation(this.latitude, this.longitude);
 
   Map<String, double> toMap() {
     return {"latitude": this.latitude, "longitude": this.longitude};
