@@ -12,15 +12,13 @@ import 'package:moengage_flutter/moe_push_service.dart';
 class MoEAndroidCore {
   MethodChannel _channel;
 
-  MoEAndroidCore(MethodChannel channel) {
-    _channel = channel;
-  }
+  MoEAndroidCore(this._channel);
 
   void enableSDKLogs() {
     _channel.invokeMethod(methodEnableSDKLogs);
   }
 
-  void trackEvent(String eventName, MoEProperties eventAttributes) {
+  void trackEvent(String eventName, MoEProperties? eventAttributes) {
     _channel.invokeMethod(
         methodTrackEvent, _getEventPayloadJSON(eventName, eventAttributes));
   }
@@ -187,7 +185,7 @@ class MoEAndroidCore {
         methodUpdateSdkState, _getUpdateSdkStatePayloadJSON(shouldEnableSdk));
   }
 
-  String _getEventPayloadJSON(String eventName, MoEProperties eventAttributes) {
+  String _getEventPayloadJSON(String eventName, MoEProperties? eventAttributes) {
     return json.encode(getEventPayload(eventName, eventAttributes));
   }
 
