@@ -1,8 +1,4 @@
-import 'dart:io';
-import 'dart:math';
-
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
 import 'dart:async';
 import 'package:moengage_flutter/moengage_flutter.dart';
 import 'package:moengage_flutter/geo_location.dart';
@@ -114,23 +110,23 @@ class _MyAppState extends State<MyApp> {
                         .addAttribute("attrName2", false)
                         .addAttribute("attrName3", 123563563)
                         .addAttribute("arrayAttr", [
-                          "str1",
-                          12.8,
-                          "str2",
-                          123,
-                          true,
-                          {"hello": "testing"}
-                        ])
+                      "str1",
+                      12.8,
+                      "str2",
+                      123,
+                      true,
+                      {"hello": "testing"}
+                    ])
                         .setNonInteractiveEvent()
                         .addAttribute(
-                            "location1", new MoEGeoLocation(12.1, 77.18))
+                        "location1", new MoEGeoLocation(12.1, 77.18))
                         .addAttribute(
-                            "location2", new MoEGeoLocation(12.2, 77.28))
+                        "location2", new MoEGeoLocation(12.2, 77.28))
                         .addAttribute(
-                            "location3", new MoEGeoLocation(12.3, 77.38))
+                        "location3", new MoEGeoLocation(12.3, 77.38))
                         .addISODateTime("dateTime1", "2019-12-02T08:26:21.170Z")
                         .addISODateTime(
-                            "dateTime2", "2019-12-06T08:26:21.170Z");
+                        "dateTime2", "2019-12-06T08:26:21.170Z");
                     _moengagePlugin.trackEvent('event flutter 01', details);
                   }),
               new ListTile(
@@ -146,33 +142,37 @@ class _MyAppState extends State<MyApp> {
                         .addAttribute("attrName2", false)
                         .addAttribute("attrName3", 123563563)
                         .addAttribute("arrayAttr", [
-                          "str1",
-                          12.8,
-                          "str2",
-                          123,
-                          true,
-                          {"hello": "testing"}
-                        ])
+                      "str1",
+                      12.8,
+                      "str2",
+                      123,
+                      true,
+                      {"hello": "testing"}
+                    ])
                         .addAttribute(
-                            "location1", new MoEGeoLocation(12.1, 77.18))
+                        "location1", new MoEGeoLocation(12.1, 77.18))
                         .addAttribute(
-                            "location2", new MoEGeoLocation(12.2, 77.28))
+                        "location2", new MoEGeoLocation(12.2, 77.28))
                         .addAttribute(
-                            "location3", new MoEGeoLocation(12.3, 77.38))
+                        "location3", new MoEGeoLocation(12.3, 77.38))
                         .addISODateTime("dateTime1", "2019-12-02T08:26:21.170Z")
                         .addISODateTime(
-                            "dateTime2", "2019-12-06T08:26:21.170Z");
+                        "dateTime2", "2019-12-06T08:26:21.170Z");
                     _moengagePlugin.trackEvent(
                         'interactive event flutter 01', details);
                   }),
               new ListTile(
                   title: Text("Track Only Event"),
                   onTap: () {
+                    _moengagePlugin.trackEvent("trackOnlyEventName", null);
+
                     _moengagePlugin.trackEvent("testEvent", MoEProperties());
                   }),
               new ListTile(
                   title: new Text("Set Unique Id"),
                   onTap: () {
+//                    _moengagePlugin.setUniqueId(null);
+
                     _moengagePlugin.setUniqueId('mobiledevs@moengage.com');
                   }),
               new ListTile(
@@ -258,7 +258,7 @@ class _MyAppState extends State<MyApp> {
               new ListTile(
                   title: Text("iOS -- Register For Push[New Instance]"),
                   onTap: () {
-                  	MoEngageFlutter newInstance = MoEngageFlutter();
+                    MoEngageFlutter newInstance = MoEngageFlutter();
                     newInstance.registerForPushNotification();
                   }),
               new ListTile(
@@ -290,6 +290,7 @@ class _MyAppState extends State<MyApp> {
                   title: Text("Android -- FCM Push Token"),
                   onTap: () {
 //                     Token passed here is just for illustration purposes. Please pass the actual token instead.
+//                    _moengagePlugin.passFCMPushToken(null);
                     _moengagePlugin.passFCMPushToken(
                         "fjt-NFxzQey7Y8mSNBig0M:APA91bGRrvQxbgebauzU4xp6yz-uQkNsPk52t1RLn5ZSZK4LTd_jpC0wGKSrI1mUHyRKgmlQbQ8r3Xt1C9aJiBCCx2F9hThJVoONSAf8fkJ31ikPkrGOYkvxcQb1s9zYtoKyCYANdZJq");
                   }),
@@ -297,8 +298,8 @@ class _MyAppState extends State<MyApp> {
                   title: Text("Android -- PushKit Push Token"),
                   onTap: () {
                     // Token passed here is just for illustration purposes. Please pass the actual token instead.
-                    _moengagePlugin.passPushKitPushToken(
-                        "IQAAAACy0T43AADshvE4JWn5zbicfxAYnljrKzjiHyUytoK-V6U0zmrjsluIB1a0oSybQlTI7_39bHJ3cix_vI6QnEx1_sT1gFULXZtCkjVn93PCdg");
+//                    _moengagePlugin.passPushKitPushToken(
+//                        "IQAAAACy0T43AADshvE4JWn5zbicfxAYnljrKzjiHyUytoK-V6U0zmrjsluIB1a0oSybQlTI7_39bHJ3cix_vI6QnEx1_sT1gFULXZtCkjVn93PCdg");
                   }),
               new ListTile(
                   title: Text("Android -- FCM Push Payload"),
@@ -310,9 +311,9 @@ class _MyAppState extends State<MyApp> {
                     pushPayload.putIfAbsent(
                         "gcm_notificationType", () => "normal notification");
                     pushPayload.putIfAbsent("gcm_alert", () => "Message");
-                    pushPayload.putIfAbsent("gcm_campaign_id", () => "1234567");
+                    pushPayload.putIfAbsent("gcm_campaign_id", () => "1234568");
                     pushPayload.putIfAbsent("gcm_activityName",
-                        () => "com.moe.pushlibrary.activities.MoEActivity");
+                            () => "com.moengage.sampleapp.MainActivity");
                     _moengagePlugin.passFCMPushPayload(pushPayload);
                   }),
               new ListTile(
@@ -350,14 +351,17 @@ class _MyAppState extends State<MyApp> {
               new ListTile(
                 title: Text("Get all messages"),
                 onTap: () async {
-                  InboxData data = await _moEngageInbox.fetchAllMessages();
-                  print("Inbox messages: " + data.toString());
-                  if (data.messages.length > 0) {
-                    _moEngageInbox.trackMessageClicked(data.messages[0]);
-                    _moEngageInbox.deleteMessage(data.messages[0]);
-                  }
-                  for (final message in data.messages) {
-                    print(message.toString());
+                  InboxData? data = await _moEngageInbox.fetchAllMessages();
+                  if (data != null && data.messages.length > 0) {
+                    InboxMessage message = data.messages[0];
+                    _moEngageInbox.trackMessageClicked(message);
+                    _moEngageInbox.deleteMessage(message);
+
+                    for (final message in data.messages) {
+                      print(message.toString());
+                    }
+                  } else {
+//                    print("Inbox messages: " + data?.toString());
                   }
                 },
               ),
