@@ -11,12 +11,11 @@ import 'constants.dart';
 class MoEAndroidInbox {
   MethodChannel _channel;
 
-  MoEAndroidInbox(MethodChannel channel) {
-    _channel = channel;
-  }
+  MoEAndroidInbox(this._channel);
 
-  Future<int> getUnClickedCount() {
-    return _channel.invokeMethod(METHOD_NAME_UN_CLICKED_COUNT);
+  Future<int> getUnClickedCount() async {
+    int count = await _channel.invokeMethod(METHOD_NAME_UN_CLICKED_COUNT);
+    return count;
   }
 
   void trackMessageClicked(InboxMessage message) {
