@@ -21,8 +21,9 @@ Map<String, dynamic> messageToMap(InboxMessage inboxMessage) {
     TEXT_CONTENT: mapFromTextContent(inboxMessage.textContent),
     ACTION: actionsListFromModel(inboxMessage.action)
   };
-  if (inboxMessage.media != null) {
-    message[MEDIA_CONTENT] = mapFromMedia(inboxMessage.media);
+  Media? media = inboxMessage.media;
+  if (media != null) {
+    message[MEDIA_CONTENT] = mapFromMedia(media);;
   }
   return message;
 }
@@ -33,9 +34,9 @@ InboxData deSerializeInboxMessages(dynamic messagesPayload) {
 }
 
 List<InboxMessage> messagesJsonToList(List<dynamic> messageArray) {
-  List<InboxMessage> messages = List<InboxMessage>();
+  List<InboxMessage> messages = [];
   for (final message in messageArray) {
-    InboxMessage inboxMessage = messageFromJson(message);
+    InboxMessage? inboxMessage = messageFromJson(message);
     if (inboxMessage != null) {
       messages.add(inboxMessage);
     }
@@ -43,7 +44,7 @@ List<InboxMessage> messagesJsonToList(List<dynamic> messageArray) {
   return messages;
 }
 
-InboxMessage messageFromJson(Map<String, dynamic> message) {
+InboxMessage? messageFromJson(Map<String, dynamic> message) {
   try {
     return InboxMessage(
         message.containsKey(ID) ? message[ID] : -1,
@@ -78,7 +79,7 @@ TextContent textContentFromMap(Map<String, dynamic> textMap) {
           : "");
 }
 
-Media mediaFromMap(Map<String, dynamic> mediaMap) {
+Media? mediaFromMap(Map<String, dynamic> mediaMap) {
   if (mediaMap.isEmpty) {
     return null;
   }
@@ -86,9 +87,9 @@ Media mediaFromMap(Map<String, dynamic> mediaMap) {
 }
 
 List<Action> actionsFromMap(List<dynamic> actions) {
-  List<Action> actionList = List<Action>();
+  List<Action> actionList = [];
   for (final action in actions) {
-    Action parsedAction = actionFromMap(action);
+    Action? parsedAction = actionFromMap(action);
     if (parsedAction != null) {
       actionList.add(parsedAction);
     }
@@ -96,13 +97,12 @@ List<Action> actionsFromMap(List<dynamic> actions) {
   return actionList;
 }
 
-Action actionFromMap(Map<String, dynamic> actionMap) {
+Action? actionFromMap(Map<String, dynamic> actionMap) {
   ActionType actionType = ActionTypeExt.fromString(actionMap[ACTION_TYPE]);
   switch (actionType) {
     case ActionType.navigation:
       return navigationActionFromMap(actionType, actionMap);
   }
-  return null;
 }
 
 NavigationAction navigationActionFromMap(
@@ -128,9 +128,9 @@ Map<String, String> mapFromMedia(Media media) {
 }
 
 List<Map<String, dynamic>> actionsListFromModel(List<Action> actions) {
-  List<Map<String, dynamic>> actionsList = List<Map<String, dynamic>>();
+  List<Map<String, dynamic>> actionsList = [];
   for (final action in actions) {
-    Map<String, dynamic> actionMap = actionToMap(action);
+    Map<String, dynamic>? actionMap = actionToMap(action);
     if (actionMap != null) {
       actionsList.add(actionMap);
     }
@@ -138,12 +138,11 @@ List<Map<String, dynamic>> actionsListFromModel(List<Action> actions) {
   return actionsList;
 }
 
-Map<String, dynamic> actionToMap(Action action) {
+Map<String, dynamic>? actionToMap(Action action) {
   switch (action.actionType) {
     case ActionType.navigation:
-      return navigationActionToMap(action);
+      return navigationActionToMap(action as NavigationAction);
   }
-  return null;
 }
 
 Map<String, dynamic> navigationActionToMap(NavigationAction navigationAction) {
